+++ sprint10/package.json
@@ -15,7 +15,9 @@
     "passport-google-oauth20": "^2.0.0",
     "passport-jwt": "^4.0.1",
     "passport-local": "^1.0.0",
-    "superstruct": "^2.0.2"
+    "path": "^0.12.7",
+    "superstruct": "^2.0.2",
+    "url": "^0.11.4"
   },
   "name": "sprint10",
   "version": "1.0.0",
