+++ sprint10/prisma/schema.prisma
@@ -64,6 +64,7 @@ model Article {
   authorId        Int?
   author          User?             @relation(fields: [authorId], references: [id], onDelete: SetNull)
   content         String
+  images          String[]
   articleComments ArticleComment[]
   createdAt       DateTime          @default(now())
   updatedAt       DateTime          @updatedAt
